+++ hydra-node/src/Hydra/Node/State.hs
@@ -15,6 +15,17 @@ import Hydra.Tx (
 
 type PendingDeposits tx = Map (TxIdType tx) (Deposit tx)
 
+data ChainPointTime = ChainPointTime
+  { currentSlot :: ChainSlot
+  -- ^ Latest chain slot as observed on chain.
+  , currentChainTime :: UTCTime
+  -- ^ Time corresponding to `currentSlot`.
+  , drift :: Natural
+  -- ^ Time difference with current system wall-clock measured in seconds
+  }
+  deriving stock (Eq, Show, Generic)
+  deriving anyclass (ToJSON, FromJSON)
+
 data NodeState tx
   = -- | Normal operation of the node where it is connected and has a recent
     -- view of the chain.
@@ -24,10 +35,7 @@ data NodeState tx
       -- ^ Pending deposits as observed on chain.
       -- TODO: could even move the chain state here (also see todo below)
       -- , chainState :: ChainStateType tx
-      , currentSlot :: ChainSlot
-      -- ^ Latest chain slot as observed on chain.
-      , currentChainTime :: UTCTime
-      -- ^ Time corresponding to `currentSlot`.
+      , chainPointTime :: ChainPointTime
       }
   | -- | Node is catching up on its view of the chain and should behave
     -- differently.
@@ -37,10 +45,7 @@ data NodeState tx
       -- ^ Pending deposits as observed on chain.
       -- TODO: could even move the chain state here (also see todo below)
       -- , chainState :: ChainStateType tx
-      , currentSlot :: ChainSlot
-      -- ^ Latest chain slot as observed on chain.
-      , currentChainTime :: UTCTime
-      -- ^ Time corresponding to `currentSlot`.
+      , chainPointTime :: ChainPointTime
       }
   deriving stock (Generic)
 
@@ -54,8 +59,15 @@ initNodeState chainState =
   NodeCatchingUp
     { headState = Idle IdleState{chainState}
     , pendingDeposits = mempty
-    , currentSlot = chainStateSlot chainState
+    , chainPointTime = initialChainPointTime chainState
+    }
+
+initialChainPointTime :: IsChainState tx => ChainStateType tx -> ChainPointTime
+initialChainPointTime chainState =
+  ChainPointTime
+    { currentSlot = chainStateSlot chainState
     , currentChainTime = initialChainTime
+    , drift = 0
     }
 
 initialChainTime :: UTCTime